[Code Health] Replace anonymous classes with lambdas in hub, init
This change applies several IDE-driven cleanups to improve readability
and code health in org.chromium.chrome.browser (hub, identity_disc,
incognito, init):

- Anonymous type can be replaced with lambda
- Anonymous type can be replaced with method reference
- Inline variable / Redundant local variable
- Redundant type cast
- Unnecessary unboxing

Bug: 541671762
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8300435
Auto-Submit: Gazal Agarwal <agazal@google.com>
Commit-Queue: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1687575}

+++ chrome/android/java/src/org/chromium/chrome/browser/hub/HubTabSwitcherMetricsRecorder.java
@@ -72,7 +72,7 @@ public class HubTabSwitcherMetricsRecorder {
         Tab previousTab = tabModel.getTabById(lastId);
         if (previousTab == null) return;
 
-        if (mPaneIdWhenShown.intValue() == currentPane.getPaneId()) {
+        if (mPaneIdWhenShown == currentPane.getPaneId()) {
             if (tab.getId() == mTabIdWhenShown) {
                 // TODO(crbug.com/40132120): Differentiate list.
                 RecordUserAction.record("MobileTabReturnedToCurrentTab.TabGrid");
